MOON-36 : split code with custom hooks
