+++ src/components/Menu/Menu.stories.jsx
@@ -13,7 +13,7 @@ storiesOf('Components|Menu', module)
     .addDecorator(withKnobs)
     .add('default', () => (
         <div style={{transform: 'scale(1)', height: '100vh'}}>
-            <Menu isDisplayed={boolean('display', true)}>
+            <Menu isDisplayed={boolean('display', true)} style={{zIndex: 10000}}>
                 <MenuItem label="Base items" variant="title"/>
                 <MenuItem label="Item1"/>
                 <MenuItem label="Item2"/>
